refactor: update loader implementation, the same as ofnil repo
