+++ src/hirag_mcp/loader/ppt_loader.py
@@ -1,5 +1,6 @@
 from langchain_community import document_loaders
 from hirag_mcp.loader.base_loader import BaseLoader
+from hirag_mcp.loader.markify_loader import markify_client
 
 
 class PowerPointLoader(BaseLoader):
@@ -9,3 +10,4 @@ class PowerPointLoader(BaseLoader):
 
     def __init__(self):
         self.loader_type = document_loaders.UnstructuredPowerPointLoader
+        self.loader_markify = markify_client
